implementacion hibernate

+++ src/mavenProyecto/PrestamosUtil.java
@@ -11,7 +11,7 @@ public class PrestamosUtil {
 		Map<Integer,Prestamo> prestamos = new HashMap<Integer, Prestamo>();
 		
 		for (Prestamo prestamo : setA) {
-			prestamos.put(prestamo.getCopia().getId(), prestamo);
+			prestamos.put((int) prestamo.getCopia().getId(), prestamo);
 		}
 
 		return prestamos;
